+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.0.0-rc.1] - 2020-02-16
+
 ### Changed
 
 * use object for option
@@ -431,7 +433,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * First release.
 
-[Unreleased]: https://github.com/shimataro/value-schema/compare/v2.2.6...HEAD
+[Unreleased]: https://github.com/shimataro/value-schema/compare/v3.0.0-rc.1...HEAD
+[3.0.0-rc.1]: https://github.com/shimataro/value-schema/compare/v2.2.6...v3.0.0-rc.1
 [2.2.6]: https://github.com/shimataro/value-schema/compare/v2.2.5...v2.2.6
 [2.2.5]: https://github.com/shimataro/value-schema/compare/v2.2.4...v2.2.5
 [2.2.4]: https://github.com/shimataro/value-schema/compare/v2.2.3...v2.2.4